Move suffstat_traits into traits
