+++ src/data/mod.rs
@@ -17,7 +17,7 @@ pub use stat::UnitPowerLawSuffStat;
 use crate::dist::{
     Bernoulli, Categorical, Gaussian, InvGamma, InvGaussian, Poisson,
 };
-use crate::suffstat_traits::{HasSuffStat, SuffStat};
+use crate::traits::{HasSuffStat, SuffStat};
 
 pub type BernoulliData<'a, X> = DataOrSuffStat<'a, X, Bernoulli>;
 pub type CategoricalData<'a, X> = DataOrSuffStat<'a, X, Categorical>;
